* main:
  fix(iOS): download beauty resources from Artifactory
  feat(iOS): update APIExample to SDK 4.6.4
  feat: support APIExample release 4.6.4
  fix(android): default audio channel profile to live broadcasting
  fix(android): align channel profile with live broadcasting

+++ iOS/APIExample/cloud_build.sh
@@ -102,28 +102,26 @@ echo "-----------------------------------"
 cd ${PROJECT_PATH}
 
 #下载美颜资源
-echo "start download bytedance resource : $bytedance_lib"
-curl -L -O "$bytedance_lib"
-unzip -o vender_bytedance_iOS.zip
-rm -f vender_bytedance_iOS.zip
-
-echo "start download sense resource : $sense_lib"
-curl -L -O "$sense_lib"
-unzip -o vender_sense_iOS.zip
+echo "start download sense resource from Artifactory"
+curl --fail --location -H "X-JFrog-Art-Api:${JFROG_API_KEY}" -o vender_sense_iOS.zip "https://artifactory-api.bj2.agoralab.co/artifactory/qa_test_data/beauty/vender_sense_iOS.zip" || exit 1
+rm -rf SenseLib
+unzip -q vender_sense_iOS.zip || exit 1
 rm -f vender_sense_iOS.zip
+test -f SenseLib/SENSEME.lic || exit 1
 
-echo "start download fu resource : $fu_lib"
-curl -L -O "$fu_lib"
-unzip -o vender_fu_iOS.zip
+echo "start download fu resource from Artifactory"
+curl --fail --location -H "X-JFrog-Art-Api:${JFROG_API_KEY}" -o vender_fu_iOS.zip "https://artifactory-api.bj2.agoralab.co/artifactory/qa_test_data/beauty/vender_fu_iOS.zip" || exit 1
+rm -rf FULib
+unzip -q vender_fu_iOS.zip || exit 1
 rm -f vender_fu_iOS.zip
+test -f FULib/authpack.h || exit 1
 
 #打开第三方播放器配置
 sed -i -e "s#\#  pod 'ijkplayer'#  pod 'ijkplayer'#g" Podfile
 
 #打开第三方美颜
-sed -i -e "s#\#  pod 'SenseLib'#  pod 'SenseLib'#g" Podfile
-sed -i -e "s#\#  pod 'bytedEffect'#  pod 'bytedEffect'#g" Podfile
-sed -i -e "s#\#  pod 'fuLib'#  pod 'fuLib'#g" Podfile
+sed -i -e "s#\#pod 'SenseLib'#pod 'SenseLib'#g" Podfile
+sed -i -e "s#\#pod 'fuLib'#pod 'fuLib'#g" Podfile
 
 echo "work space: $WORKSPACE"
 echo "project path: $PROJECT_PATH"
